// src/api/stetsom/model/productCatalogResponse.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { ProductCardItem } from "./productCardItem";

export interface ProductCatalogResponse {
  items: ProductCardItem[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}